semigrp: Adds unittests, docs and fix exceptions
Changes FullTransformationMonoid to full_transformation_monoid. Edits
index.ipynb to include this.

Adds documentation for the Semigroup class and full_transformation_monoid
function.

Adds exception to __init__ in the Semigroup class, when called with a Python
object, before passing to PythonElementNC, checks the element can be
multiplied by itself.

Adds error messages for exceptions.

Adds unittests for semigroups.

Disables pylint len-as-condition, so that a list can be passed, as
well as a number of arguments not in a list.